Test whether vector query with genuine yields genuine, hr yields hr
In order to test my theory (that genuine reviews and hr reviews occupy different regions in a vector space), store the training data in a vector db and test to see if searching for a genuine review (from the test set) yields a genuine review in the result; and likewise, test to see if searching for an HR-written review (from the test set) yields a hr-written review in the result.

It does, and I didn't expect it to work :) Elhamdulillah!